Refactor database interaction endpoints into a separate module
